extract interface from event logger
